feat: add createTranslation helper for app dir (#1150)
